consolidate 3 recipe brain structs into RecipeBrain; fix PR #2138 findings
Task 1: Merge RecipeDecideBrain, RecipeOrientBrain, and
RecipeEngineerLifecycleBrain into a single RecipeBrain struct
parameterised by (recipe_filename, adapter_tag). Delete the 3 redundant
files. Update brains.rs wiring. Keep type aliases for backward compat.

Task 2a: session_builder.rs — resolve_agent_binary() now logs config
errors via tracing::warn! instead of swallowing them into None.

Task 2b: runtime_config.rs — default_state_root() now warns when HOME
env var is unset and the hardcoded fallback is used.

Task 2c: Delete 4 overcooked doc files (config-driven-subprocess-env,
configure-llm-provider, disk-health-api, subprocess-env-propagation)
and clean cross-links in 5 remaining docs.

140 tests in recipe_brain.rs (37 TDD + 103 migrated from deleted files).
4951 total tests pass, clippy clean, fmt clean.

Closes #2132
